Db query code optimization
